Added kv_store and unit test framework

+++ compile.sh
@@ -1,9 +1,12 @@
-rm -rf ./build
+rebuild "$@"
+
+if [[ "$rebuild" == "true" ]]; then
+  rm -rf ./build
+fi
 mkdir -p ./build
 
 cd ./build
 cmake ..
 make
 
-cd ..
-./build/kv_cli
+ctest --rerun-failed --output-on-failure
